+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 
 ## [15.1.0.1]
 ### Added
+- TLS enabled ECDSA by default for ESP8266
 
 ### Breaking Changed